fix rewrite multi sub group ncfile error

+++ cdm/src/main/java/ucar/nc2/iosp/netcdf3/N3header.java
@@ -682,8 +682,8 @@ public class N3header {
         if (largeFile)
           raf.writeLong(pos);
         else {
-          if (pos > Integer.MAX_VALUE)
-            throw new IllegalArgumentException("Variable starting pos="+pos+" may not exceed "+ Integer.MAX_VALUE);          
+          if (pos > Long.MAX_VALUE)
+            throw new IllegalArgumentException("Variable starting pos="+pos+" may not exceed "+ Long.MAX_VALUE);
           raf.writeInt((int) pos);
         }
